Refactor, move loading and streaming code into subdirectories
