+++ python/src/scippneutron/__init__.py
@@ -5,10 +5,7 @@
 
 # flake8: noqa
 
-from ._scippneutron import __version__
-from ._scippneutron import convert
-from ._scippneutron import position, source_position, sample_position, incident_beam, scattered_beam, Ltotal, L1, L2, two_theta
 from .mantid import from_mantid, to_mantid, load, fit
 from .instrument_view import instrument_view
-from .load_nexus import load_nexus, load_nexus_json
-from .data_stream import data_stream, start_stream
+from .file_loading.load_nexus import load_nexus, load_nexus_json
+from .data_streaming.data_stream import data_stream, start_stream
