+++ python/src/scippneutron/data_streaming/_consumer.py
