+++ python/src/scippneutron/file_loading/_common.py
