+++ python/src/scippneutron/file_loading/_nexus.py
@@ -1,5 +1,5 @@
-from ._loading_json_nexus import LoadFromJson
-from ._loading_hdf5_nexus import LoadFromHdf5
+from ._json_nexus import LoadFromJson
+from ._hdf5_nexus import LoadFromHdf5
 from typing import Union, Dict
 import h5py
 import scipp as sc
